refactor: stm32f4 can模块重构完成

+++ src/librm/hal/can_interface.h
@@ -43,7 +43,7 @@ namespace rm::hal {
 
 struct CanMsg {
   std::array<u8, 8> data;
-  u32 rx_std_id;
+  u16 rx_std_id;
   u32 dlc;
 };
